backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy==2.0.23
PostgreSQL==0.1.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
web3==6.11.3
cryptography==41.0.8
bcrypt==4.1.2
marshmallow==3.20.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
